+++ spotinst_sdk/test/test_aws_elastigroup.py
@@ -23,7 +23,7 @@ class AwsElastigroupTestCase(unittest.TestCase):
 
     @staticmethod
     def load_group_json():
-        with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'test_lib/group.json')) as group_json:
+        with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'test_lib/input/group.json')) as group_json:
             return json.load(group_json)