+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.39] - Unreleased
+### Added
+ - kubernetes api calls
+ - b/g deployment api calls
+ - beanstalk api calls
+ - stateful api calls
+ - process and scaling policy api calls
+ - deployment api calls
+ - instance api calls
+ - cost api calls
+ - other misc api calls
+
 ## [1.0.39] - 2018-10-04
 ### Updated
  - update_elastigroup() had error with the send_put() request arguments